Splitted the file into two components, and runned the --production
